refactor(core): move lifecycle test module into tests.rs submodule
mod.rs was 1520 lines (over the 1500 GodFile threshold) because the
inline #[cfg(test)] mod tests block was left in it. Moved the test
module into agent_task_lifecycle/tests.rs (1472 lines) so mod.rs is a
48-line re-export aggregation point, keeping every submodule under the
structural threshold.
